requirements.txt
absl-py==0.13.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
distro==1.6.0
docopt==0.6.2
google-auth==1.31.0
google-auth-oauthlib==0.4.4
grpcio==1.38.0
idna==2.10
imageio==2.9.0
importlib-metadata==4.5.0
joblib==1.0.1
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.54
packaging==21.0
pandas==1.1.5
Pillow==8.2.0
pipreqs==0.4.11
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-build==0.12.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.1.0
seaborn==0.11.1
six==1.16.0
sklearn==0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.2
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2020.9.3
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
yarg==0.1.9
zipp==3.4.1
